structure changed, added user and film endpoints
