+++ src/swaggerDocumentation/swaggerCinemaHalls.ts
@@ -1,0 +1,13 @@
+export const cinemaHallModel = {
+    type: "object",
+    properties: {
+        name: {
+            type: "string",
+            required: "true",
+        },
+        numberOfSeats: {
+            type: "number",
+            required: "true",
+        },
+    },
+}
